# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.PermissionPolicy:
    properties:
      policy_document:
        type: string
      policy_id:
        type: string
      policy_name:
        type: string
      policy_type:
        $ref: '#/definitions/domain.PolicyType'
      provider:
        $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider'
    type: object
  domain.PolicyType:
    enum:
    - system
    - custom
    type: string
    x-enum-varnames:
    - PolicyTypeSystem
    - PolicyTypeCustom
...
  github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider:
    enum:
    - aliyun
    - aws
    - azure
    - tencent
    - huawei
    - volcano
    - volcengine
    type: string
    x-enum-comments:
      CloudProviderAWS: Amazon Web Services
      CloudProviderAliyun: 阿里云
      CloudProviderAzure: Microsoft Azure
      CloudProviderHuawei: 华为云
      CloudProviderTencent: 腾讯云
      CloudProviderVolcano: 火山云 (别名)
      CloudProviderVolcengine: 火山引擎
    x-enum-descriptions:
    - 阿里云
    - Amazon Web Services
    - Microsoft Azure
    - 腾讯云
    - 华为云
    - 火山云 (别名)
    - 火山引擎
    x-enum-varnames:
    - CloudProviderAliyun
    - CloudProviderAWS
    - CloudProviderAzure
    - CloudProviderTencent
    - CloudProviderHuawei
    - CloudProviderVolcano
    - CloudProviderVolcengine
...
  service.GroupSyncResult:
    properties:
      created_groups:
        type: integer
      failed_groups:
        type: integer
      failed_members:
        type: integer
      synced_members:
        type: integer
      total_groups:
        type: integer
      total_members:
        type: integer
      updated_groups:
        type: integer
    type: object
...
  web.CreateUserGroupVO:
    properties:
      cloud_platforms:
        items:
          $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider'
        minItems: 1
        type: array
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      policies:
        items:
          $ref: '#/definitions/domain.PermissionPolicy'
        type: array
      tenant_id:
        type: string
    required:
    - cloud_platforms
    - name
    - tenant_id
    type: object
...
  web.GenerateAuditReportVO:
    properties:
      end_time:
        type: string
      start_time:
        type: string
      tenant_id:
        type: string
    required:
    - end_time
    - start_time
    - tenant_id
    type: object
...
  web.PageResult:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  web.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  web.UpdatePoliciesVO:
    properties:
      policies:
        items:
          $ref: '#/definitions/domain.PermissionPolicy'
        minItems: 1
        type: array
    required:
    - policies
    type: object
...
  web.UpdateUserGroupVO:
    properties:
      cloud_platforms:
        items:
          $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider'
        type: array
      description:
        type: string
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/domain.PermissionPolicy'
        type: array
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业云资产管理服务 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: E-CAM Service API
  version: "1.0"
paths:
  /api/v1/cam/iam/audit/logs:
    get:
      parameters:
      - description: 操作类型
        in: query
        name: operation_type
        type: string
      - description: 操作人ID
        in: query
        name: operator_id
        type: string
      - description: 目标类型
        in: query
        name: target_type
        type: string
      - description: 云平台
        in: query
        name: cloud_platform
        type: string
      - description: 租户ID
        in: query
        name: tenant_id
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.PageResult'
      summary: 查询审计日志列表
      tags:
      - 审计日志管理
  /api/v1/cam/iam/audit/logs/export:
    get:
      parameters:
      - description: 操作类型
        in: query
        name: operation_type
        type: string
      - description: 操作人ID
        in: query
        name: operator_id
        type: string
      - description: 目标类型
        in: query
        name: target_type
        type: string
      - description: 云平台
        in: query
        name: cloud_platform
        type: string
      - description: 租户ID
        in: query
        name: tenant_id
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - description: 导出格式(csv/json)
        in: query
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 导出审计日志
      tags:
      - 审计日志管理
  /api/v1/cam/iam/audit/reports:
    post:
      consumes:
      - application/json
      parameters:
      - description: 生成审计报告请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.GenerateAuditReportVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 生成审计报告
      tags:
      - 审计日志管理
  /api/v1/cam/iam/groups:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.PageResult'
      summary: 查询用户组列表
      tags:
      - 用户组管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 创建用户组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateUserGroupVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 创建用户组
      tags:
      - 用户组管理
  /api/v1/cam/iam/groups/{id}:
    delete:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 删除用户组
      tags:
      - 用户组管理
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户组详情
      tags:
      - 用户组管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新用户组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.UpdateUserGroupVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 更新用户组信息
      tags:
      - 用户组管理
  /api/v1/cam/iam/groups/{id}/members:
    get:
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户组成员列表
      tags:
      - 用户组管理
  /api/v1/cam/iam/groups/{id}/policies:
    put:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 用户组ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新权限策略请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.UpdatePoliciesVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 更新用户组的权限策略
      tags:
      - 用户组管理
  /api/v1/cam/iam/groups/sync:
    post:
      description: |-
        从云平台同步用户组信息，包括用户组基本信息、权限策略和用户组成员。支持阿里云、腾讯云等多云平台。
        同步结果包含：用户组统计（总数、新建、更新、失败）和成员统计（总数、同步成功、失败）
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 云账号ID
        in: query
        name: cloud_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 同步结果统计
          schema:
            allOf:
            - $ref: '#/definitions/web.Result'
            - properties:
                data:
                  $ref: '#/definitions/service.GroupSyncResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/web.Result'
        "500":
          description: 同步失败
          schema:
            $ref: '#/definitions/web.Result'
      summary: 同步云平台用户组及其成员
      tags:
      - 用户组管理
  /api/v1/cam/iam/permissions/groups/{group_id}:
    get:
      parameters:
      - description: 用户组ID
        in: path
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户组的所有权限
      tags:
      - 权限管理
...
schemes:
- http
- https
swagger: "2.0"